Refactor chatManager
