+++ Assets/AgoraEngine/RtmDemo/MessageDisplay.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e84db2b8f83c22942b97521605f8d0c8
+guid: f2821b8e26f7942a7985b65c466fda06
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
