+++ Assets/AgoraEngine/RtmDemo/RtmChatManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fada1dae731e4492a83d78e079c3324b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
